implimented comments from previous pull request
