+++ __tests__/app.test.js
@@ -64,7 +64,6 @@ describe("GET /api/articles/:article_id", () => {
             topic: expect.any(String),
             created_at: expect.any(String),
             votes: expect.any(Number),
-            comment_count: 11,
           })
         );
       });
@@ -85,6 +84,19 @@ describe("GET /api/articles/:article_id", () => {
       .then(({ body }) => {
         const { msg } = body;
         expect(msg).toBe("Invalid ID!");
+      });
+  });
+  test("Should return an article object with comment count", () => {
+    return request(app)
+      .get("/api/articles/1")
+      .then(({ body }) => {
+        const { article } = body;
+        expect(article).toBeInstanceOf(Object);
+        expect(article).toEqual(
+          expect.objectContaining({
+            comment_count: 11,
+          })
+        );
       });
   });
 });
