+++ controllers/articles.controllers.js
@@ -1,0 +1,31 @@
+const {
+  retrieveArticle,
+  changeArticle,
+  commentCount,
+} = require("../models/articles.models");
+
+exports.getArticle = (req, res, next) => {
+  const id = req.params;
+  retrieveArticle(id.article_id)
+    .then((article) => {
+      res.status(200).send({ article });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
+
+exports.updateArticle = (req, res, next) => {
+  const articleInfo = {
+    inc_votes: req.body.inc_votes,
+    article_id: req.params.article_id,
+  };
+
+  changeArticle(articleInfo)
+    .then((article) => {
+      res.status(200).send({ article });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
