+++ controllers/users.controllers.js
@@ -1,0 +1,7 @@
+const { retrieveUsers } = require("../models/users.models.js");
+
+exports.getUsers = (req, res, next) => {
+  retrieveUsers().then((users) => {
+    res.status(200).send({ users });
+  });
+};
