+++ models/topics.models.js
@@ -1,0 +1,13 @@
+const db = require("../db/connection");
+
+exports.retrieveTopics = () => {
+  return db.query("SELECT * FROM topics").then(({ rows }) => {
+    return rows;
+  });
+};
+
+exports.retrieveUsers = () => {
+  return db.query("SELECT * FROM users").then(({ rows }) => {
+    return rows;
+  });
+};
