+++ models/users.models.js
@@ -1,0 +1,7 @@
+const db = require("../db/connection");
+
+exports.retrieveUsers = () => {
+  return db.query("SELECT * FROM users").then(({ rows }) => {
+    return rows;
+  });
+};
